Add comments and move Card to resources
