+++ src/Backend/Resources/Resource.php
@@ -10,7 +10,6 @@ use Illuminate\Support\Facades\Validator as ValidatorFactory;
 use Illuminate\Support\Str;
 use Illuminate\Validation\ValidationException;
 use Illuminate\Validation\Validator;
-use Oscer\Cms\Backend\Elements\Card;
 use Oscer\Cms\Backend\Resources\Fields\Field;
 
 abstract class Resource implements \JsonSerializable
@@ -47,36 +46,47 @@ abstract class Resource implements \JsonSerializable
      */
     protected function resolveFields()
     {
+        // if this method was not executed
         if ($this->fields->isEmpty()) {
+            // temporary variable to get a flat collection of fields
             $fields = new Collection();
+            // Iterate over the collection returned by the fields method
             $this->fields()->each(function ($field) use ($fields) {
+                // These instances can also be cards
                 if ($field instanceof Card) {
+                    // if it is a card we need all its fields...
                     foreach ($field->fields as $fieldInCard) {
+                        // We need to resolve the values for the fields from the resource model.
+                        $fieldInCard->resolve($this->resourceModel);
+                        // add them to our temporary collection
                         $fields->add($fieldInCard);
                     }
-
-                    return;
                 }
-                if (! $field->card) {
-                    $field->card = 'default';
+                elseif ($field instanceof Field) {
+                    // the rest should be field instances...
+                    if (!$field->card) {
+                        // Assign the default card to all fields without card assignment
+                        $field->card = 'default';
+                    }
+                    $field->resolve($this->resourceModel);
+                    // add them to our temporary collection
+                    $fields->add($field);
                 }
-                $fields->add($field);
             });
 
-            $this->fields = $fields->map(function (Field $field) {
-                $field->value = $field->resolve($this->resourceModel);
-
-                return $field;
-            });
+            // We need to resolve the values for the fields from the resource model.
+            // The result will be assigned to the fields property of this resource.
+            $this->fields = $fields;
         }
 
+        // Return the fields
         return $this->fields;
     }
 
     protected function filteredFields(Request $request): Collection
     {
         return $this->resolveFields()->filter(function (Field $field) use ($request) {
-            return ! $field->shouldBeRemoved($request);
+            return !$field->shouldBeRemoved($request);
         });
     }
